truths_lies_generator_main.cc: optimize all truths collection
Instead of storing a single vector of sets (allTruthsPerGenerator), we can
create a vector of sets (truthsSetPerGenerator) and a vector of vectors
(truthsVectorPerGenerator). In each collection, we use shared pointers to the
statements, thereby reducing memory usage. This then allows us to do quick set
containment lookups as well as indexed lookups.